+++ dlt/destinations/impl/bigquery/configuration.py
@@ -20,7 +20,7 @@ class BigQueryClientConfiguration(DestinationClientDwhWithStagingConfiguration):
         60.0  # how long to retry the operation in case of error, the backoff 60 s.
     )
     loading_api: str = "default"
-    batch_size: int = 0
+    batch_size: int = 500
 
     __config_gen_annotations__: ClassVar[List[str]] = ["location"]
