+++ dlt/destinations/job_impl.py
@@ -1,10 +1,20 @@
+from abc import ABC, abstractmethod
 import os
 import tempfile  # noqa: 251
+from typing import Dict, Iterable, List
 
-from dlt.common.storages import FileStorage
-
+from dlt.common import json
 from dlt.common.destination.reference import NewLoadJob, FollowupJob, TLoadJobState, LoadJob
-from dlt.common.storages.load_storage import ParsedLoadJobFileName
+from dlt.common.schema import Schema, TTableSchema
+from dlt.common.storages import FileStorage
+from dlt.common.typing import TDataItems
+
+from dlt.destinations.impl.destination.configuration import (
+    GenericDestinationClientConfiguration,
+    TDestinationCallable,
+)
+
+from dlt.pipeline.current import commit_load_package_state
 
 
 class EmptyLoadJobWithoutFollowup(LoadJob):
@@ -60,3 +70,109 @@ class NewReferenceJob(NewLoadJobImpl):
         with open(file_path, "r+", encoding="utf-8") as f:
             # Reading from a file
             return f.read()
+
+
+class DestinationLoadJob(LoadJob, ABC):
+    def __init__(
+        self,
+        table: TTableSchema,
+        file_path: str,
+        config: GenericDestinationClientConfiguration,
+        schema: Schema,
+        destination_state: Dict[str, int],
+        destination_callable: TDestinationCallable,
+        skipped_columns: List[str],
+    ) -> None:
+        super().__init__(FileStorage.get_file_name_from_file_path(file_path))
+        self._file_path = file_path
+        self._config = config
+        self._table = table
+        self._schema = schema
+        # we create pre_resolved callable here
+        self._callable = destination_callable
+        self._state: TLoadJobState = "running"
+        self._storage_id = f"{self._parsed_file_name.table_name}.{self._parsed_file_name.file_id}"
+        self.skipped_columns = skipped_columns
+        try:
+            if self._config.batch_size == 0:
+                # on batch size zero we only call the callable with the filename
+                self.call_callable_with_items(self._file_path)
+            else:
+                current_index = destination_state.get(self._storage_id, 0)
+                for batch in self.run(current_index):
+                    self.call_callable_with_items(batch)
+                    current_index += len(batch)
+                    destination_state[self._storage_id] = current_index
+
+            self._state = "completed"
+        except Exception as e:
+            self._state = "retry"
+            raise e
+        finally:
+            # save progress
+            commit_load_package_state()
+
+    @abstractmethod
+    def run(self, start_index: int) -> Iterable[TDataItems]:
+        pass
+
+    def call_callable_with_items(self, items: TDataItems) -> None:
+        if not items:
+            return
+        # call callable
+        self._callable(items, self._table)
+
+    def state(self) -> TLoadJobState:
+        return self._state
+
+    def exception(self) -> str:
+        raise NotImplementedError()
+
+
+class DestinationParquetLoadJob(DestinationLoadJob):
+    def run(self, start_index: int) -> Iterable[TDataItems]:
+        # stream items
+        from dlt.common.libs.pyarrow import pyarrow
+
+        # guard against changed batch size after restart of loadjob
+        assert (
+            start_index % self._config.batch_size
+        ) == 0, "Batch size was changed during processing of one load package"
+
+        # on record batches we cannot drop columns, we need to
+        # select the ones we want to keep
+        keep_columns = list(self._table["columns"].keys())
+        start_batch = start_index / self._config.batch_size
+        with pyarrow.parquet.ParquetFile(self._file_path) as reader:
+            for record_batch in reader.iter_batches(
+                batch_size=self._config.batch_size, columns=keep_columns
+            ):
+                if start_batch > 0:
+                    start_batch -= 1
+                    continue
+                yield record_batch
+
+
+class DestinationJsonlLoadJob(DestinationLoadJob):
+    def run(self, start_index: int) -> Iterable[TDataItems]:
+        current_batch: TDataItems = []
+
+        # stream items
+        with FileStorage.open_zipsafe_ro(self._file_path) as f:
+            encoded_json = json.typed_loads(f.read())
+            if isinstance(encoded_json, dict):
+                encoded_json = [encoded_json]
+
+            for item in encoded_json:
+                # find correct start position
+                if start_index > 0:
+                    start_index -= 1
+                    continue
+                # skip internal columns
+                for column in self.skipped_columns:
+                    item.pop(column, None)
+                current_batch.append(item)
+                if len(current_batch) == self._config.batch_size:
+                    yield current_batch
+                    current_batch = []
+            yield current_batch
