+++ tests/load/pipeline/test_bigquery.py
@@ -7,6 +7,7 @@ from dlt.common.schema import Schema
 from dlt.common.schema.typing import TTableSchema
 from dlt.common.typing import TDataItems
 from google.cloud import bigquery
+from dlt.destinations.impl.bigquery.bigquery_adapter import bigquery_adapter
 
 from tests.pipeline.utils import assert_load_info
 from tests.load.pipeline.utils import destinations_configs, DestinationTestConfiguration
@@ -46,6 +47,19 @@ def test_bigquery_numeric_types(destination_config: DestinationTestConfiguration
 
 def test_bigquery_streaming_insert():
     pipe = dlt.pipeline(destination="bigquery")
-    pack = pipe.run([{"field": 1}, {"field": 2}], table_name="test_streaming_items")
+    pack = pipe.run([{"field1": 1, "field2": 2}], table_name="test_streaming_items")
+
+    assert_load_info(pack)
+
+
+def test_bigquery_adapter_streaming_insert():
+    @dlt.resource
+    def test_resource():
+        yield {"field1": 1, "field2": 2}
+
+    bigquery_adapter(test_resource, insert_api="streaming")
+
+    pipe = dlt.pipeline(destination="bigquery")
+    pack = pipe.run(test_resource, table_name="test_streaming_items")
 
     assert_load_info(pack)
